30-06-2025

+++ src/main/java/in/co/rays/Proj04/model/RoleModel.java
@@ -163,7 +163,7 @@ public class RoleModel {
 
 	}
 
-	public RoleBean findByPk(int id) {
+	public RoleBean findByPk(long id) {
 
 		Connection conn = null;
 		RoleBean bean = null;
